Update FastUploadAllocator to handle large allocation requests
The FastUploadAllocator would have integer overflow in the case of large
allocation requests paired with an existing, relatively large
`m_spaceUsed`. This was particularly noticeable when we were passing
UNKNOWN_VERTEX_COUNT as the size to allocate.

Known issue with this: If size is >= (1 << 31), CielToClosestPowerOfTwo
fails and m_size is set to 0. In Debug mode this would cause an assert,
but in release it just fails. Not sure if we should handle that case and
return UINT_MAX, or if that should throw an exception since 32 bit apps
shouldn't be trying to use that much memory anyway.