Refactor 9on12draw.inl into separate components
Static functions are difficult to test in isolation, especially when
they depend on other static functions, and even more so when those
functions are in the same file. By splitting this inl file into
logically distinct chuncks we can more easily test these functions in
isolation. We can also still have the 9on12 pch include them as normal,
so 9on12 should be none the wiser.