Prepare 9on12Device for mocking
The constructor now has fewer responsibilities. Although it still does
more initialization than I would like, as long as we don't call the
`Init` method I think we can work around it. This only allows us to Mock
9on12Device. If we want to add tests for 9on12Device we will need to do
further refactoring to be able to mock out it's components as well.