Claim support for Dither
Dithering only really made sense when rendering to reduced bit-depth
formats (565, 5551, etc), but DX9 era hardware generally had 8 bit
precision in pixel shaders anyway so it was usually a noop. DXVK also
chose to simply ignore dithering, so we should be fine to claim support
but not actually do anything.