Add Tracking for drawing PreTransformedVertices
In D3D9 it was possible to pre transform and light vertices before
drawing. This process also allowed culling, so the GPU wouldn't cull in
this case. The DDIs that correspond to this for 9on12 are the
DrawPrimitive2 and DrawIndexedPrimitive2 calls, so we need to keep track
of which "mode" we are in so that we don't cull/clip when we shouldn't.