[DX9]Cities: Skylines - Invisible water
Cities used INTZ surface to read in the shader that rendered the water patches to
calculate depth from water surface to the backing z surface. Cities
bound the INTZ as SRV and kept the same INTZ surface bound for depth
testing and the depth writes enabled. Dx9on12 checked for SRV resource
being bound for DSV and ZWriteEnable in ResolveDeferredState() from
9on12PixelStage.cpp, the combination would cause the resource to be
unbound. Dx9on12 was doing the right thing, it should be application's
responsibility to remove the SRV.

Reviewed upon INTZ documents, it appeared that when using INTZ as shader
resources and keeping it bound as depth required the ZWriteEnable to be
false. The proposed fix detected for INTZ resources bound as SRV and
DSV, instead of hiding SRV, disable ZWriteEnable and
continued to bind the resource as SRV. At the end of drawcall, restored
ZWriteEnable state.