Update DDI
We will need a few new DDIs in order to implement the new Present path.
Unfortunately, the existing CreateDeviceArgs doesn't have a concept of
versioning. We need to add that and try to limit the impact of garbage
data when reading the new interfaceVersion field (since it didn't exist
before we'll be reading off the end of the struct passed by the d3d9
runtime if it is out of sync with 9on12).

We also need a new way to handle fetching the privateDDITable as we are
now adding DDIs out of sync with the runtime. Old runtimes will still call
GetPrivateDDITable which will copy the original 27 DDIs, but newer runtimes
will be able to call GetPrivateDDITableVersioned which allows the caller to
pass the expected size of the private ddi table struct. Since we don't
change the order of the DDIs between versions, this is a safe way to
make sure that the runtime always gets the full ddi table that it
expects without overwriting regions of memory owned by the runtime.