Update DDI
We will need a few new DDIs in order to implement the new Present path.
Unfortunately, the existing CreateDeviceArgs doesn't have a concept of
versioning. This requires the addition of a new OpenAdapter2_Private DDI
so that we can use a new version of the CreateDeviceArgs struct that
contains versioning info.

We also need a new way to handle fetching the privateDDITable as we are
now adding DDIs out of sync with the runtime. Old runtimes will still call
GetPrivateDDITable which will copy the original 27 DDIs, but newer runtimes
will be able to call GetPrivateDDITableVersioned which allows the caller to
pass the expected size of the private ddi table struct. Since we don't
change the order of the DDIs between versions, this is a safe way to
make sure that the runtime always gets the full ddi table that it
expects without overwriting regions of memory owned by the runtime.