Change InputBuffer::m_sizeInBytes default to UINT32_MAX
m_sizeInBytes is only used for bounds checking, but we don't always
have a valid size to go off of, particularly for user memory buffers.

When possible, we will take advantage of it for the bounds check, but if
it isn't available we should default to trusting the app to pass valid
data. UINT32_MAX will always be greater than the attempted upload size,
so this is a safe default in this case.