[Translate_RCP] Swap FLT_MAX for src0
This is technically not to spec, as we should be using FLT_MAX here.
However, since CSGO (d3d9 version) was found to expect non-FLT_MAX
values when doing an rcp of 0, we're going to keep the out of spec
approach.

We do need the special handling for when src0 has modifiers on it though
as movc doesn't support that.