Allow passing any non-zero value to enable alpha test or fog
9on12 previously expected apps to pass an explicit TRUE (1) value for
enabling alphaTest or fog. This is technically to spec, but some apps
such as S.T.A.L.K.E.R. Shadow of Chernobyl would pass 0xFF instead with
the intent to enable.

Went ahead and applied this change to FogEnable too for consistency
since I wouldn't be surprised to hit that in the future.

Fixes #86